补充了单独查询的有关接口 修正了一些返回错误

+++ RatingSystemServer/src/main/java/com/example/service/UserService.java
@@ -31,5 +31,5 @@ public interface UserService extends IService<User> {
 
     User getUserByUsername(String username);
 
-    R getUserById(int id);
+    R getUserById(long id);
 }
